Refactor scripts to battle-move
This introduces a new class, BattleMove, available as `battle.move`,
for most of the functions that were previously in `data/scripts.ts`.

This makes it so that "go to definition" will now work correctly for
functions previously in scripts; we no longer need UnimplementedError,
and there's now a clean conceptual separation between `battle` and
`battle-move` (whereas the previous distinction between `battle` and
`scripts` was basically nonexistent).